Search Button Edits
This commit fixes and improves on the search button in the header. It fixes the border-left length. As you can see in the first example in the image below, the length goes past it's parent container.

The other changes in this commit include a border on the search button :hover that is the npmred color darkened a bit. I also fixed some odd behavior of the button width becoming shorter when :hover (ed) on. Seemed a bit odd as it made the button look like it was jumping right a little bit.

Lastly, I recreated the logo in svg format. See issue #73. Note: you may not be able to tell a difference unless you have a retina display.