Adds test case for packages like benchmark that previously failed to be sanitized.
The major issue with the `benchmark` package was that its readme wasn't at the top level. As a result, it wasn't being elevated to the top level and then sanitized before being served to the registry. This test ensures that even older packages (circa 2012, which is before we forced all packages to have readmes at the top level) are sanitized.

* Takes registry data for test package (`benchmark`)
* Ensures mocked methods serves the correct registry info
* Tests that the package is sanitized to html from markdown