Fixed issue #499: handle mixed-case package names
Don't attempt to validate package names before fetching, because
the rules have changed along the way. Only validate if a package
is not found, so we can vary the message shown to the viewer.
(Now-invalid names do not get any encouragement to publish with
that name.)

Added `JSONStream` to the list of replicated packages for verification.