Promise.catch(err).then() considered incorrect.
The correct order of operations here is
`Promise.then(normal-flow).catch(error);`
If you catch an error, return from that function, then pass the
result along to a then(), whatever's in the then will be executed.

Caught this because a misconfigured service was causing all star
operations to error, which then caused a double-reply error in www.