Registry relational queries are now cached.
The package "model" now uses the cache for all package data gets.
It drops any cached data for a specific package on update. There
is now a convenience function for generating the predictable package
Request options, so it is guaranteed to look the same every time.

Added promisified versions of `cache.get()` and `cache.drop()`
for convenience, called `cache.getP()` and `cache.dropP()`.
Added unit tests for the promisified wrappers.

Adjusted two package "model" tests to look for the regularized
error messages that the cache generates on 404s.

All tests pass. This of course means success is assured.