Overzealous regex was logging people out
If a package had 'static' in the name, the url wasn't being
authenticated against. We'd like to not go through the process of
AUTHing requests for static files, but of course don't want to log
people out because they look at a package that has 'static' as part of
the title.

Fixes #1032