Give all images max-width 100%
Ok, so this affects all images on the site. It's a pretty standard
reset, but that doesn't mean it doesn't warrant a discussion. Do we
prefer that people have the ability to have images that are too big?
And that blow our layouts out?

If so, this isn't the solution for us.

That being said, I've done some pretty heavy spot testing and haven't
found a place where we lose out yet.

If anybody has any spots, let me know.

Fixes #1060