Add support for an Org's "full name"
fullname is an optional attribute for an organization that is currently
stored in the user-acl `resource` hstore (as `human_name`) but will eventually become a
column on that model.

Since it is stored as `human_name` on the user-acl, we finesse it a
little along the way.

This updates tests, fixtures, handlers, and the agent