Depend on npm@3
Funny story, this one

We install `nemo` which installs `npmi` which installs `npm@2` as a
convenience for installing selenium.

This means that when we install newww with npm 3, we get npm 2 in
`node_modules/.bin/`, and therefore in our `PATH`, meaning all
invocations of `npm` from an `npm` script will use npm 2, not npm 3.

To make matters worse, npm 2 really can't handle the tree that npm 3 has
laid down, so things can really get screwy.

So we depend on npm just to make sure a sensible version is present in
the PATH at all times.